+++ Runtime/LLM.cs
@@ -65,7 +65,7 @@ namespace LLMUnity
         /// }
         /// \endcode
         /// </summary>
-        [LLMAdvanced] public bool asynchronousStartup = false;
+        [LLMAdvanced] public bool asynchronousStartup = true;
         /// <summary> select to not destroy the LLM GameObject when loading a new Scene. </summary>
         [LLMAdvanced] public bool dontDestroyOnLoad = true;
         /// <summary> the path of the model being used (relative to the Assets/StreamingAssets folder).
@@ -79,13 +79,15 @@ namespace LLMUnity
         [ModelAdvanced] public int contextSize = 0;
         /// <summary> Batch size for prompt processing. </summary>
         [ModelAdvanced] public int batchSize = 512;
+        /// <summary> a base prompt to use as a base for all LLMCharacter objects </summary>
+        [TextArea(5, 10), ChatAdvanced] public string basePrompt = "";
+
         /// <summary> Boolean set to true if the server has started and is ready to receive requests, false otherwise. </summary>
         public bool started { get; protected set; } = false;
         /// <summary> Boolean set to true if the server has failed to start. </summary>
         public bool failed { get; protected set; } = false;
 
         /// \cond HIDE
-        public string slotSaveDir;
         public int SelectedModel = 0;
         [HideInInspector] public float modelProgress = 1;
         [HideInInspector] public float modelCopyProgress = 1;
@@ -174,13 +176,13 @@ namespace LLMUnity
             // Start the LLM server in a cross-platform way
             if (model == "")
             {
-                Debug.LogError("No model file provided!");
+                LLMUnitySetup.LogError("No model file provided!");
                 return null;
             }
             string modelPath = LLMUnitySetup.GetAssetPath(model);
             if (!File.Exists(modelPath))
             {
-                Debug.LogError($"File {modelPath} not found!");
+                LLMUnitySetup.LogError($"File {modelPath} not found!");
                 return null;
             }
             string loraPath = "";
@@ -189,7 +191,7 @@ namespace LLMUnity
                 loraPath = LLMUnitySetup.GetAssetPath(lora);
                 if (!File.Exists(loraPath))
                 {
-                    Debug.LogError($"File {loraPath} not found!");
+                    LLMUnitySetup.LogError($"File {loraPath} not found!");
                     return null;
                 }
             }
@@ -199,7 +201,6 @@ namespace LLMUnity
             if (remote) arguments += $" --port {port} --host 0.0.0.0";
             if (numThreads > 0) arguments += $" -t {numThreads}";
             if (loraPath != "") arguments += $" --lora \"{loraPath}\"";
-            arguments += $" --slot-save-path \"{slotSaveDir}\"";
             arguments += $" -ngl {numGPULayers}";
             return arguments;
         }
@@ -211,15 +212,15 @@ namespace LLMUnity
         public async void Awake()
         {
             if (!enabled) return;
-            slotSaveDir = Application.persistentDataPath;
             if (asynchronousStartup) await Task.Run(() => StartLLMServer());
             else StartLLMServer();
             if (dontDestroyOnLoad) DontDestroyOnLoad(transform.root.gameObject);
+            if (basePrompt != "") await SetBasePrompt(basePrompt);
         }
 
         private void SetupLogging()
         {
-            logStreamWrapper = ConstructStreamWrapper(Debug.LogWarning, true);
+            logStreamWrapper = ConstructStreamWrapper(LLMUnitySetup.LogWarning, true);
             llmlib?.Logging(logStreamWrapper.GetStringWrapper());
         }
 
@@ -237,7 +238,7 @@ namespace LLMUnity
             string arguments = GetLlamaccpArguments();
             if (arguments == null) return;
             bool useGPU = numGPULayers > 0;
-            Debug.Log($"Server command: {arguments}");
+            LLMUnitySetup.Log($"Server command: {arguments}");
 
             foreach (string arch in LLMLib.PossibleArchitectures(useGPU))
             {
@@ -246,7 +247,7 @@ namespace LLMUnity
                 {
                     InitLib(arch);
                     InitServer(arguments);
-                    Debug.Log($"Using architecture: {arch}");
+                    LLMUnitySetup.Log($"Using architecture: {arch}");
                     break;
                 }
                 catch (LLMException e)
@@ -258,16 +259,16 @@ namespace LLMUnity
                 {
                     error = $"{e.GetType()}: {e.Message}";
                 }
-                Debug.Log($"Tried architecture: {arch}, " + error);
+                LLMUnitySetup.Log($"Tried architecture: {arch}, " + error);
             }
             if (llmlib == null)
             {
-                Debug.LogError("LLM service couldn't be created");
+                LLMUnitySetup.LogError("LLM service couldn't be created");
                 failed = true;
                 return;
             }
             StartService();
-            Debug.Log("LLM service created");
+            LLMUnitySetup.Log("LLM service created");
         }
 
         private void InitLib(string arch)
@@ -343,7 +344,7 @@ namespace LLMUnity
             else if (!started) error = "LLM service not started";
             if (error != null)
             {
-                Debug.LogError(error);
+                LLMUnitySetup.LogError(error);
                 throw new Exception(error);
             }
         }
@@ -358,12 +359,12 @@ namespace LLMUnity
             string message = $"LLM {status}: {result}";
             if (status > 0)
             {
-                if (log) Debug.LogError(message);
+                if (log) LLMUnitySetup.LogError(message);
                 throw new LLMException(message, status);
             }
             else if (status < 0)
             {
-                if (log) Debug.LogWarning(message);
+                if (log) LLMUnitySetup.LogWarning(message);
             }
         }
 
@@ -432,6 +433,7 @@ namespace LLMUnity
         public async Task<string> Completion(string json, Callback<string> streamCallback = null)
         {
             AssertStarted();
+            if (streamCallback == null) streamCallback = (string s) => {};
             StreamWrapper streamWrapper = ConstructStreamWrapper(streamCallback);
             await Task.Run(() => llmlib.LLM_Completion(LLMObject, json, streamWrapper.GetStringWrapper()));
             if (!started) return null;
@@ -440,6 +442,13 @@ namespace LLMUnity
             DestroyStreamWrapper(streamWrapper);
             CheckLLMStatus();
             return result;
+        }
+
+        public async Task SetBasePrompt(string base_prompt)
+        {
+            AssertStarted();
+            SystemPromptRequest request = new SystemPromptRequest(){system_prompt = base_prompt, prompt = " ", n_predict = 0};
+            await Completion(JsonUtility.ToJson(request));
         }
 
         /// <summary>
@@ -479,7 +488,7 @@ namespace LLMUnity
             }
             catch (Exception e)
             {
-                Debug.LogError(e.Message);
+                LLMUnitySetup.LogError(e.Message);
             }
         }