+++ Runtime/LLMInterface.cs
@@ -40,6 +40,14 @@ namespace LLMUnity
     }
 
     [Serializable]
+    public struct SystemPromptRequest
+    {
+        public string prompt;
+        public string system_prompt;
+        public int n_predict;
+    }
+
+    [Serializable]
     public struct ChatResult
     {
         public int id_slot;
@@ -96,7 +104,7 @@ namespace LLMUnity
     {
         public int id_slot;
         public string action;
-        public string filename;
+        public string filepath;
     }
 
     [Serializable]