+++ Samples~/KnowledgeBaseGame/KnowledgeBaseGame.cs
@@ -166,13 +166,19 @@ namespace LLMUnitySamples
             AIReplyComplete();
         }
 
+        public void ExitGame()
+        {
+            Debug.Log("Exit button clicked");
+            Application.Quit();
+        }
+
         bool onValidateWarning = true;
         void OnValidate()
         {
             if (onValidateWarning)
             {
                 if (embedding.SelectedOption == 0) Debug.LogWarning($"Please select a model in the {embedding.gameObject.name} GameObject!");
-                if (llmCharacter.llm.model == "") Debug.LogWarning($"Please select a model in the {llmCharacter.llm.gameObject.name} GameObject!");
+                if (!llmCharacter.remote && llmCharacter.llm != null && llmCharacter.llm.model == "") Debug.LogWarning($"Please select a model in the {llmCharacter.llm.gameObject.name} GameObject!");
                 onValidateWarning = false;
             }
         }