+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## v2.0.3
+#### 🚀 Features
+
+- Add LLM selector in Inspector mode (PR: #182)
+- Allow to save chat history at custom path (PR: #179)
+- Use asynchronous startup by default (PR: #186)
+- Assign LLM if not set according to the scene and hierarchy (PR: #187)
+- Allow to set log level (PR: #189)
+- Allow to add callback functions for error messages (PR: #190)
+- Allow to set a LLM base prompt for all LLMCharacter objects (PR: #192)
+
+#### 🐛 Fixes
+
+- set higher priority for mac build with Accelerate than without (PR: #180)
+- Fix duplicate bos warning in llama3 (PR: #188)
+
+
 ## v2.0.2
 #### 🐛 Fixes